+++ src/CMakeLists.txt
@@ -5,7 +5,7 @@
 add_subdirectory(agents)
 add_subdirectory(backend)
 add_subdirectory(common)
-add_subdirectory(maliput)
+add_subdirectory(roads)
 add_subdirectory(gen)
 add_subdirectory(protobuf)
 add_subdirectory(systems)
